requirements.txt
uvicorn==0.15.0
starlette==0.14.2
numpy==1.21.2
transformers==4.10.0
fastapi==0.68.1
Pillow==8.3.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
python-multipart==0.0.5